MR-title: TSK2.2.1B：后端：火车餐及外卖系统基本模型与接口
Created-by: tsfksj
Author-id: 1216290
MR-id: 1342873
Commit-by: SwiftScout
Merged-by: hid_h_9i3x6cs6hzgor
Description: merge "tsk-2-2-1-B-meal-sys" into "dev"
fix(TakeawayDishDataConverter): close #69035038 外卖餐品的Model中，shop_id应当暂时可空,
fix(TrainRepository):
修复`get_trains`中错误使用了train_type_id、seat_type_id、train_id,
fix: 为加载大于1GB的JSON，修改最大请求体大小,
docs(backend): 更新部署方式,
feat: 实现加载火车餐、外卖数据功能,
feat: 新增save_raw_takeaway、save_raw_dish方法,
feat(OrderStatusConsumer): 将消费者加入OrderStatusConsumerService中,
feat: 实现DishOrderStatusConsumer、TakeawayOrderStatusConsumer,
feat: 实现DishBookingService、TakeawayBookingService,
feat(Order): 新增`set_status`方法,
feat: 实现TakeawayShopRepository,
feat: 实现DishRepository,
refactor(db): 修改dish, takeaway_dish, takeaway_shop的外键为integer类型

See merge request: c30259be6f9d41a2987593e524170eff/SwiftJourney!38

+++ backend/base/src/models/dish.rs
@@ -7,7 +7,7 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
-    pub train_id: i64,
+    pub train_id: i32,
     pub r#type: String,
     pub time: String,
     pub name: String,
